pyproject.toml
[project]
name = "rnoise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.4",
    "perlin-noise>=1.13",
    "pillow>=11.1.0",
]
